refactor: 分割しすぎを抑えてEditorViewへ集約 / Refactor reduce over-splitting by consolidating into EditorView
